requirements.txt
asgiref==3.2.3
boto3==1.16.43
botocore==1.19.43
Django==3.0.3
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.11.0
docutils==0.14
gunicorn==20.0.4
jmespath==0.10.0
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2019.3
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.26.2
whitenoise==5.2.0